requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
distro==1.8.0
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
oauthlib==3.2.2
openai==1.3.4
Pillow==10.1.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
whitenoise==6.6.0
